Fix ar9x compile/boot (#7102)
* view_main: fix compile on atsam

* AR9X: fix broken/missing boardInit() since 0ec34fd72c5d1ab356c9129aac52985466005654, readd rtcInit to boardInit (on AR9X)

+++ radio/src/gui/128x64/view_main.cpp
@@ -186,7 +186,7 @@ void drawTimerWithMode(coord_t x, coord_t y, uint8_t index)
     const uint8_t negative = (timerState.val<0 ? BLINK | INVERS : 0);
     LcdFlags att = RIGHT | DBLSIZE;
     if (timerState.val < 60 * 60) { // display MM:SS
-      div_t qr = div(abs(timerState.val), 60);
+      div_t qr = div((int)abs(timerState.val), 60);
       lcdDrawNumber(x - 5, y, qr.rem, att | LEADING0 | negative, 2);
       lcdDrawText(lcdLastLeftPos, y, ":", att|BLINK | negative);
       lcdDrawNumber(lcdLastLeftPos, y, qr.quot, att | negative);
@@ -194,7 +194,7 @@ void drawTimerWithMode(coord_t x, coord_t y, uint8_t index)
         lcdDrawText(lcdLastLeftPos, y, "-", att | negative);
     }
     else if (timerState.val < (99 * 60 * 60) + (59 * 60)) { // display HHhMM
-      div_t qr = div(abs(timerState.val) / 60, 60);
+      div_t qr = div((int)(abs(timerState.val) / 60), 60);
       lcdDrawNumber(x - 5, y, qr.rem, att | LEADING0, 2);
       lcdDrawText(lcdLastLeftPos, y, "h", att);
       lcdDrawNumber(lcdLastLeftPos, y, qr.quot, att);
